backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
asyncmy
aiomysql
pymysql
pydantic[email]
email-validator
bcrypt
python-jose[cryptography]
requests
python-multipart
google-generativeai
razorpay
reportlab
pandas
openpyxl
python-dotenv
slowapi
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
